Move validate_transaction_error_trace_attributes into validator directory
