+++ tests/agent_features/test_ignore_expected_errors.py
@@ -19,11 +19,20 @@ from testing_support.fixtures import (
     validate_error_event_attributes_outside_transaction,
     validate_error_event_sample_data,
     validate_error_trace_attributes_outside_transaction,
+)
+from testing_support.validators.validate_time_metrics_outside_transaction import (
+    validate_time_metrics_outside_transaction,
+)
+from testing_support.validators.validate_transaction_error_trace_attributes import (
     validate_transaction_error_trace_attributes,
 )
-from testing_support.validators.validate_time_metrics_outside_transaction import validate_time_metrics_outside_transaction
-from testing_support.validators.validate_transaction_errors import validate_transaction_errors
-from testing_support.validators.validate_transaction_metrics import validate_transaction_metrics
+from testing_support.validators.validate_transaction_errors import (
+    validate_transaction_errors,
+)
+from testing_support.validators.validate_transaction_metrics import (
+    validate_transaction_metrics,
+)
+
 from newrelic.api.application import application_instance
 from newrelic.api.background_task import background_task
 from newrelic.api.time_trace import notice_error
@@ -36,12 +45,8 @@ _error_message = "Test error message."
 # Settings presets
 
 # Error classes settings
-expected_runtime_error_settings = {
-    "error_collector.expected_classes": [_runtime_error_name]
-}
-ignore_runtime_error_settings = {
-    "error_collector.ignore_classes": [_runtime_error_name]
-}
+expected_runtime_error_settings = {"error_collector.expected_classes": [_runtime_error_name]}
+ignore_runtime_error_settings = {"error_collector.ignore_classes": [_runtime_error_name]}
 
 # Status code settings
 expected_status_code_settings = {"error_collector.expected_status_codes": [418]}
@@ -140,9 +145,7 @@ override_expected_matrix = (True, False, None)
 
 @pytest.mark.parametrize("settings,expected", error_trace_settings_matrix)
 @pytest.mark.parametrize("override_expected", override_expected_matrix)
-def test_error_trace_attributes_inside_transaction(
-    settings, expected, override_expected
-):
+def test_error_trace_attributes_inside_transaction(settings, expected, override_expected):
     expected = override_expected if override_expected is not None else expected
 
     error_trace_attributes = {
@@ -164,9 +167,7 @@ def test_error_trace_attributes_inside_transaction(
 
 @pytest.mark.parametrize("settings,expected", error_trace_settings_matrix)
 @pytest.mark.parametrize("override_expected", override_expected_matrix)
-def test_error_trace_attributes_outside_transaction(
-    settings, expected, override_expected
-):
+def test_error_trace_attributes_outside_transaction(settings, expected, override_expected):
     expected = override_expected if override_expected is not None else expected
 
     error_trace_attributes = {
@@ -181,9 +182,7 @@ def test_error_trace_attributes_outside_transaction(
     }
 
     @reset_core_stats_engine()
-    @validate_error_trace_attributes_outside_transaction(
-        _runtime_error_name, exact_attrs=error_trace_attributes
-    )
+    @validate_error_trace_attributes_outside_transaction(_runtime_error_name, exact_attrs=error_trace_attributes)
     @override_application_settings(settings)
     def _test():
         exercise(override_expected)
@@ -205,9 +204,7 @@ def test_error_metrics_inside_transaction(expected):
         ("ErrorsExpected/all", expected_metrics_count),
     ]
 
-    @validate_transaction_metrics(
-        "test", background_task=True, rollup_metrics=metrics_payload
-    )
+    @validate_transaction_metrics("test", background_task=True, rollup_metrics=metrics_payload)
     @background_task(name="test")
     def _test():
         exercise(expected)
