+++ tests/agent_features/test_wsgi_attributes.py
@@ -17,9 +17,11 @@ from testing_support.fixtures import (
     dt_enabled,
     override_application_settings,
     validate_error_event_attributes,
-    validate_transaction_error_trace_attributes,
 )
 from testing_support.sample_applications import fully_featured_app
+from testing_support.validators.validate_transaction_error_trace_attributes import (
+    validate_transaction_error_trace_attributes,
+)
 from testing_support.validators.validate_transaction_event_attributes import (
     validate_transaction_event_attributes,
 )
